+++ R/compute_occurrences.R
@@ -4,6 +4,7 @@
 compute_occurrences <- function(.query){
   switch(pour("atlas", "region"),
          "Austria" = compute_occurrences_la_direct(.query),
+         "France" = compute_occurrences_la_direct(.query),
          "United Kingdom" = compute_occurrences_la_direct(.query),
          "Global" = compute_occurrences_gbif(.query),
          compute_occurrences_la(.query))