+++ tests/testthat/test-international-France.R
@@ -157,10 +157,14 @@ test_that("atlas_occurrences works for France", {
   skip_if_offline()
   galah_config(atlas = "France",
                email = "ala4r@ala.org.au")
-  x <- galah_call() |>
+  base_query <- galah_call() |>
     filter(year <= 1950) |>
-    identify("Vulpes vulpes") |>
-    select(species, year) |>
+    identify("Vulpes vulpes")
+  counts <- base_query |>
+    count() |>
+    collect()
+  occ_collapse <-  base_query |>
+    # select(species, year) |>
     collapse()
   skip_if(inherits(occ_collapse, "try-error"), message = "API not available")
   expect_s3_class(occ_collapse, "query")
@@ -168,7 +172,7 @@ test_that("atlas_occurrences works for France", {
                c("type", "url", "headers", "filter"))
   expect_equal(occ_collapse$type, "data/occurrences")
   # compute
-  occ_compute <- compute(occ_collapse) # fails here
+  occ_compute <- compute(occ_collapse)
   expect_s3_class(occ_compute, "computed_query")
   # collect
   occ <- collect(occ_compute) |>
@@ -176,7 +180,7 @@ test_that("atlas_occurrences works for France", {
   skip_if(inherits(occ_compute, "try-error"), message = "API not available")
   expect_equal(nrow(occ), counts$count[1])
   expect_s3_class(occ, c("tbl_df", "tbl", "data.frame"))
-  expect_equal(ncol(occ), 2)
+  expect_equal(ncol(occ), 8)
   unlink("temp", recursive = TRUE)
 })