+++ NEWS.md
@@ -1,3 +1,15 @@
+# galah 2.0.3
+
+### Minor improvements
+* Support media downloads for most atlases (Brazil, Guatemala, Spain, Sweden and Portugal) in addition to ALA
+* Support data profiles for Sweden and Spain
+* Update to new APIs from Austria (#231)
+* Species downloads (via `atlas_species()`) now work for Sweden, France, and Spain (#234)
+* `select()` now works for species downloads (i.e. via `atlas_species()`; #227)
+* reorganise help files for improved clarity
+* Swap to new ALA APIs (issue no.)
+* Fix bugs in filter, group by etc not recognising fields (FIX MEDIA FIRST)
+
 # galah 2.0.2
 
 ### Minor improvements